# specs/1-predictive-maintenance/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Twin Platform API
  description: Predictive maintenance API for SME Digital Twin Platform
  version: 1.0.0
  contact:
    name: Digital Twin Platform Support
    email: support@digitaltwin.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.digitaltwin.com
    description: Production server
paths:
  /api/machines:
    get:
      summary: List all machines
      description: Retrieve a paginated list of all machines in the system
      tags:
        - Machine Management
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/MachineType'
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Machine'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new machine
      description: Add a new machine to the monitoring system
      tags:
        - Machine Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMachineRequest'
      responses:
        '201':
          description: Machine created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/machines/{id}:
    get:
      summary: Get machine details
      description: Retrieve detailed information about a specific machine
      tags:
        - Machine Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update machine
      description: Update machine configuration and details
      tags:
        - Machine Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMachineRequest'
      responses:
        '200':
          description: Machine updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete machine
      description: Remove a machine from the system
      tags:
        - Machine Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Machine deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Machine:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Production Line Motor 1"
        type:
          $ref: '#/components/schemas/MachineType'
        configuration:
          type: object
          example:
            powerRating: 150
            ratedSpeed: 1800
            operatingTemperature: 75
            vibrationThreshold: 5.0
        location:
          type: string
          example: "Production Area A"
        status:
          $ref: '#/components/schemas/MachineStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          example: true
    CreateMachineRequest:
      type: object
      required:
        - name
        - type
        - location
      properties:
        name:
          type: string
          maxLength: 200
        type:
          $ref: '#/components/schemas/MachineType'
        configuration:
          type: object
        location:
          type: string
          maxLength: 500
    UpdateMachineRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        configuration:
          type: object
        location:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/MachineStatus'
        isActive:
          type: boolean
    MachineType:
      type: integer
      enum:
        - 1  # Motor
        - 2  # Pump
        - 3  # Compressor
        - 4  # Gearbox
        - 5  # Bearing
    MachineStatus:
      type: integer
      enum:
        - 1  # Online
        - 2  # Offline
        - 3  # Maintenance
        - 4  # Error
        - 5  # Degraded
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
                example: "RESOURCE_NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
                example: "INTERNAL_SERVER_ERROR"
              traceId:
                type: string
                format: uuid
...
security:
  - BearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT